Extracted (part of) the authorization using config files.
