+++ src/Lithnet.Laps.Web/Lithnet.Laps.Web/Auth/IAuthService.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.DirectoryServices.AccountManagement;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -13,13 +14,13 @@ namespace Lithnet.Laps.Web.Auth
         /// access the password of the computer with name
         /// <paramref name="computerName"/>
         /// </summary>
-        /// <param name="userName">name of the user</param>
+        /// <param name="currentUser">a user. FIXME: this should just be the user name.</param>
         /// <param name="computerName">name of the computer</param>
         /// <param name="target">Target section in the web.config-file.
         /// FIXME: This shouldn't be in this interface. But I can't leave it out
         /// yet, because the code figuring out the target has way too many dependencies.
         /// </param>
-        /// <returns><c>true</c> or <c>false</c> accordingly</returns>
-        bool CanAccessPassword(string userName, string computerName, TargetElement target = null);
+        /// <returns>An <see cref="AuthResponse"/> object.</returns>
+        AuthResponse CanAccessPassword(UserPrincipal user, string computerName, TargetElement target = null);
     }
 }
